Added comfirm messages

+++ src/javascripts/events/classEvents.js
@@ -1,7 +1,7 @@
 import showAddClassesForm from '../components/classes/showAddClassesForm';
 import showAllClasses from '../components/classes/showAllClasses';
 import showEditClassForm from '../components/classes/showEditClassForm';
-import { deleteClass, updateClass } from '../helpers/classHelper';
+import { deleteClass, getSingleClass, updateClass } from '../helpers/classHelper';
 
 const handleUpdateClass = (firebaseKey) => {
   const classInfo = {
@@ -12,7 +12,11 @@ const handleUpdateClass = (firebaseKey) => {
 };
 
 const handleDeleteClass = (firebaseKey) => {
-  deleteClass(firebaseKey).then(showAllClasses);
+  getSingleClass(firebaseKey).then((classInfo) => {
+    // eslint-disable-next-line no-alert
+    const isConfirmed = window.confirm(`Do you want to delete the "${classInfo.name}" class?`);
+    if (isConfirmed) deleteClass(firebaseKey).then(showAllClasses);
+  });
 };
 
 const classEvents = (eventId) => {
